refactor(cactus-web): moving tests to the Alert.test.tsx file
CACTUS-759
